EC-597 Moved integration tests out of unit test classes
