+++ Comments.Test/UnitTests/ConsultationService.Tests.cs
@@ -43,42 +43,6 @@ namespace Comments.Test.UnitTests
             viewModel.Comments.Single().CommentText.ShouldBe(commentText);
         }
 
-        [Fact]
-        public async void CommentsQuestionsAndAnswers_CanBeRead()
-        {
-            // Arrange
-            ResetDatabase();
-            var sourceURI = "consultations://./consultation/1/document/1/chapter/introduction";
-            var commentText = Guid.NewGuid().ToString();
-            var questionText = Guid.NewGuid().ToString();
-            var answerText = Guid.NewGuid().ToString();
-            var createdByUserId = Guid.Empty;
-            var authenticateService = new FakeAuthenticateService(authenticated: true);
-
-            AddCommentsAndQuestionsAndAnswers(sourceURI, commentText, questionText, answerText, createdByUserId, (int)StatusName.Draft, _context);
-	        //var submitService = new SubmitService(_context, _fakeUserService, _consultationService);
-			var commentService = new CommentService(_context, _fakeUserService, authenticateService, _consultationService);
-
-            // Act    
-            var viewModel = commentService.GetCommentsAndQuestions(sourceURI);
-
-            //Assert
-            viewModel.Comments.Single().CommentText.ShouldBe(commentText);
-            var question = viewModel.Questions.Single();
-            question.QuestionText.ShouldBe(questionText);
-            question.Answers.Single().AnswerText.ShouldBe(answerText);
-
-
-            //sample code:
-            
-            var resp = await _client.GetAsync($"/consultations/api/Comments?sourceURI={WebUtility.UrlEncode(sourceURI)}");
-            resp.EnsureSuccessStatusCode();
-            var responseString = await resp.Content.ReadAsStringAsync();
-            var anotherQuestion = JsonConvert.DeserializeObject<CommentsAndQuestions>(responseString);
-            anotherQuestion.Questions.Single().Answers.Single().AnswerText.ShouldBe(answerText);
-            
-        }
-
 	    [Theory]
 		//document page tests
 		[InlineData(null, "gid-ng10107", "html-content", "/guidance/indevelopment/gid-ng10107/consultation/html-content", null)] //regular indev project
